refactor(#428): Consolidate duplicate SIAC/POAC/PAAC NJK templates into a single shared handler
Three identical NJK templates for siac, poac, and paac weekly hearing lists
are replaced by a single siac-poac-paac-weekly-hearing-list template in the lib.
The three individual page controllers are replaced by one handler that routes
by listTypeId (28=SIAC, 29=POAC, 30=PAAC), following the same pattern used
for FTT RPT.
