--- apps/web/src/pages/(list-types)/poac-weekly-hearing-list/poac-weekly-hearing-list.njk
@@ -1,75 +1,0 @@
-{% extends "layouts/base-template.njk" %}
-
-{% block page_content %}
-<div class="govuk-grid-row">
-  <div class="govuk-grid-column-full">
-
-    <h1 class="govuk-heading-l" id="top">{{ header.listTitle }}</h1>
-
-    <p class="govuk-body">
-      <a href="{{ t.factLinkUrl }}" class="govuk-link">{{ t.factLinkText }}</a> {{ t.factAdditionalText }}
-    </p>
-
-    <p class="govuk-body govuk-!-font-weight-bold govuk-!-margin-bottom-1">{{ t.listForWeekCommencing }} {{ header.weekCommencingDate }}</p>
-    <p class="govuk-body">{{ t.lastUpdated }} {{ header.lastUpdatedDate }} {{ t.at }} {{ header.lastUpdatedTime }}</p>
-
-    <details class="govuk-details govuk-!-margin-top-6" data-module="govuk-details" open>
-      <summary class="govuk-details__summary">
-        <span class="govuk-details__summary-text">
-          {{ t.importantInformationTitle }}
-        </span>
-      </summary>
-      <div class="govuk-details__text">
-        <p class="govuk-body">{{ t.importantInformationText }}</p>
-        <p class="govuk-body">
-          <a href="{{ t.importantInformationLinkUrl }}" class="govuk-link">{{ t.importantInformationLinkText }}</a>
-        </p>
-      </div>
-    </details>
-
-    <div class="govuk-form-group govuk-!-margin-top-6">
-      <h2 class="govuk-heading-s">{{ t.searchCasesTitle }}</h2>
-      <label class="govuk-label govuk-visually-hidden" for="case-search-input">
-        {{ t.searchCasesLabel }}
-      </label>
-      <input class="govuk-input govuk-!-width-one-half" id="case-search-input" name="search" type="text" aria-label="{{ t.searchCasesLabel }}">
-    </div>
-
-    <div id="hearings-table-container">
-      <table class="govuk-table" id="hearings-table" role="table" aria-label="{{ header.listTitle }}">
-        <thead class="govuk-table__head">
-          <tr class="govuk-table__row">
-            <th scope="col" class="govuk-table__header">{{ t.tableHeaders.date }}</th>
-            <th scope="col" class="govuk-table__header">{{ t.tableHeaders.time }}</th>
-            <th scope="col" class="govuk-table__header">{{ t.tableHeaders.appellant }}</th>
-            <th scope="col" class="govuk-table__header">{{ t.tableHeaders.caseReferenceNumber }}</th>
-            <th scope="col" class="govuk-table__header">{{ t.tableHeaders.hearingType }}</th>
-            <th scope="col" class="govuk-table__header">{{ t.tableHeaders.courtroom }}</th>
-            <th scope="col" class="govuk-table__header">{{ t.tableHeaders.additionalInformation }}</th>
-          </tr>
-        </thead>
-        <tbody class="govuk-table__body">
-          {% for hearing in hearings %}
-            <tr class="govuk-table__row">
-              <td class="govuk-table__cell">{{ hearing.date }}</td>
-              <td class="govuk-table__cell">{{ hearing.time }}</td>
-              <td class="govuk-table__cell">{{ hearing.appellant }}</td>
-              <td class="govuk-table__cell">{{ hearing.caseReferenceNumber }}</td>
-              <td class="govuk-table__cell">{{ hearing.hearingType }}</td>
-              <td class="govuk-table__cell">{{ hearing.courtroom }}</td>
-              <td class="govuk-table__cell">{{ hearing.additionalInformation }}</td>
-            </tr>
-          {% endfor %}
-        </tbody>
-      </table>
-    </div>
-
-    <p class="govuk-body-s govuk-!-margin-top-6">{{ t.dataSource }}: {{ dataSource }}</p>
-
-    <div class="govuk-!-margin-top-6">
-      <a href="#top" class="govuk-link">{{ t.backToTop }}</a>
-    </div>
-
-  </div>
-</div>
-{% endblock %}
