+++ libs/list-types/siac-poac-paac-weekly-hearing-list/package.json
@@ -13,8 +13,9 @@
     }
   },
   "scripts": {
-    "build": "tsc && yarn build:nunjucks && yarn build:schemas",
+    "build": "tsc && yarn build:nunjucks && yarn build:views && yarn build:schemas",
     "build:nunjucks": "mkdir -p dist/pdf && cd src/pdf && find . -name '*.njk' -exec sh -c 'mkdir -p ../../dist/pdf/$(dirname {}) && cp {} ../../dist/pdf/{}' \\;",
+    "build:views": "mkdir -p dist/views && find src/views -name '*.njk' -exec cp {} dist/views/ \\;",
     "build:schemas": "mkdir -p dist/schemas && cp src/schemas/*.json dist/schemas/",
     "dev": "tsc --watch",
     "test": "vitest run",
